moved pages.tsx, started extracting types and constants
