+++ packages/wrangler/src/__tests__/pages.test.ts
@@ -9,8 +9,8 @@ import {
 import { mockConsoleMethods } from "./helpers/mock-console";
 import { runInTempDir } from "./helpers/run-in-tmp";
 import { runWrangler } from "./helpers/run-wrangler";
-import type { Deployment, Project, UploadPayloadFile } from "../pages";
 import type { FormData, RequestInit } from "undici";
+import {Deployment, Project, UploadPayloadFile} from "../pages/types";
 
 // Asserting within mock responses get swallowed, so run them out-of-band
 const outOfBandTests: (() => void)[] = [];
